refactor PageStrategy
